requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
babel==2.17.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
discord.py==2.6.4
docutils==0.21.2
frozenlist==1.8.0
google-auth==2.41.1
google-genai==1.43.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
libsql-client==0.3.1
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
propcache==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
python-dotenv==1.1.1
pyturso==0.2.1
requests==2.32.5
roman-numerals-py==3.1.0
rsa==4.9.1
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
